refactor: createUser with validations
