--- src/api/userClient/userClient.types.ts
@@ -1,9 +1,0 @@
-export type CreateUserClient = {
-  id?: number;
-  name: string;
-  lastname: string;
-  email: string;
-  avatar: string;
-  password: string;
-  rol_id: number;
-};
