+++ src/api/users/users.service.ts
@@ -1,23 +1,41 @@
-import { PrismaClient } from "@prisma/client";
-import { usersCreateData , updateUser } from './users.type';
-import { hashPassword } from '../../auth/utils/bcrypt'
+import { PrismaClient } from '@prisma/client';
+import { CreateUser, updateUser } from './users.type';
+import { hashPassword } from '../../auth/utils/bcrypt';
 
 const prisma = new PrismaClient();
 
-export async function createUser(input: usersCreateData) {
-
+export async function createUser(input: CreateUser) {
+  //Add client in Users Table
   const hashedPassword = await hashPassword(input.password);
 
-  const data = {
-    ...input,
-    password: hashedPassword
-  }
   const user = await prisma.users.create({
-    data
-  })
+    data: {
+      name: input.name,
+      lastname: input.lastname,
+      email: input.email,
+      avatar: input.avatar,
+      password: hashedPassword,
+    },
+  });
 
-  return user;
+  //Add client in UsersByRol Table
+  const userRoles = input.rol_id;
 
+  userRoles.forEach(async (rolId, index) => {
+    await prisma.userByRole.create({
+      data: {
+        rolId,
+        usersId: user.id,
+      },
+    });
+  });
+
+  const newUser = {
+    ...user,
+    rol_id: userRoles,
+  };
+
+  return newUser;
 }
 
 export async function getAllUser() {
@@ -29,8 +47,8 @@ export async function getUserById(id: number) {
   const user = await prisma.users.findUnique({
     where: {
       id,
-    }
-  })
+    },
+  });
   return user;
 }
 
@@ -42,22 +60,19 @@ export async function deleteUser(id: number) {
     data: {
       isActive: false,
     },
-  })
+  });
   return updateUser;
 }
 
-export async function updateUser(
-  id: number,
-  data: updateUser
-  ) {
-console.log(id)
-console.log(data)
+export async function updateUser(id: number, data: updateUser) {
+  console.log(id);
+  console.log(data);
   const updateUser = await prisma.users.update({
     where: {
       id,
     },
-    data
-  })
+    data,
+  });
   return updateUser;
 }
